quant_project/pyproject.toml
[project]
name = "quant-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.132.0",
    "feedparser>=6.0.12",
    "html5lib>=1.1",
    "ipykernel>=7.2.0",
    "lightgbm>=4.6.0",
    "lxml>=6.0.2",
    "numpy>=2.4.2",
    "optuna>=4.7.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "praw>=7.8.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.47",
    "statsmodels>=0.14.6",
    "streamlit>=1.54.0",
    "ta>=0.11.0",
    "uvicorn>=0.41.0",
    "vadersentiment>=3.3.2",
    "vectorbt>=0.28.4",
    "xgboost>=3.2.0",
    "yfinance>=1.2.0",
]
